fix: register shared Inertia props at provider level (web#1530)
Move the shared-data bag out of HandleInertiaRequests::share() into
Inertia::share() in WebServiceProvider::boot(), mirroring the existing
provider-level setRootView().

The shared props (images/user/sidebar/locale/translations/flash/errors/
activeSidebarElement) were only ever applied by the web-group middleware.
The Pest browser in-process server doesn't run provider-pushed group
middleware, so those props were absent and DarkSidebar crashed reading
usePage().props.images.logo — the authenticated dashboard mounted an empty
#app. Registering them at the provider guarantees they apply whether or not
the middleware runs; the real app is unaffected (props identical).

activeSidebarElement becomes a per-response closure (request()->route()) so
it still resolves per request from the shared bag.
